pyproject.toml
[project]
name = "ai-verse"
version = "0.1.0"
description = "A basic AI chatbot using strands and textual."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "langchain",
    "langchain-community",
    "langchain-google-genai>=4.1.2",
    "langchain-ollama",
    "langgraph>=1.0.5",
    "praw>=7.8.1",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "textual",
]
